+++ gridsome/lib/__tests__/__fixtures__/project-basic/src/pages/About.vue
@@ -1,5 +1,6 @@
 <template>
   <Layout>
     <h1>{{ $context.title }}</h1>
+    <g-image src="../assets/test.png" immediate />
   </Layout>
 </template>