+++ gridsome/lib/graphql/filters/query.js
@@ -5,14 +5,16 @@ const listRefOpsMap = {
   '$in': '$refListIn',
   '$nin': '$refListNin',
   '$eq': '$refListEq',
-  '$ne': '$refListNe'
+  '$ne': '$refListNe',
+  '$exists': '$refListExists'
 }
 
 const refOpsMap = {
   '$in': '$refIn',
   '$nin': '$refNin',
   '$eq': '$refEq',
-  '$ne': '$refNe'
+  '$ne': '$refNe',
+  '$exists': '$refExists'
 }
 
 function toFilterArgs (filter, typeComposer, currentKey = '') {
@@ -31,7 +33,7 @@ function toFilterArgs (filter, typeComposer, currentKey = '') {
     const extensions = typeComposer.getFieldExtensions(key)
 
     // TODO: remove this workaround
-    if (extensions.isDeprecatedNodeReference) {
+    if (extensions.isInferredReference) {
       currentKey += '.id'
     }