+++ gridsome/lib/server/middlewares/assets.js
@@ -23,7 +23,7 @@ module.exports = ({ context, config, assets }) => {
       return decodeURIComponent(value)
     })
 
-    const asset = assets.get(key)
+    const asset = assets.get(key || req.url)
 
     if (!asset) return res.sendStatus(404)
 
@@ -36,8 +36,9 @@ module.exports = ({ context, config, assets }) => {
         res.header('Expires', '-1')
       }
 
+      res.header('Accept-Ranges', 'bytes')
       res.contentType(ext)
-      res.end(buffer, 'binary')
+      res.end(buffer)
     }
 
     try {
@@ -48,7 +49,7 @@ module.exports = ({ context, config, assets }) => {
 
         if (!fs.existsSync(destPath)) {
           await worker.processImage({
-            backgroundColor: config.images.backgroundColor,
+            imagesConfig: config.images,
             width: asset.width,
             height: asset.height,
             filePath,