+++ gridsome/lib/webpack/createBaseConfig.js
@@ -236,9 +236,13 @@ module.exports = (app, { isProd, isServer }) => {
       }])
   }
 
+  // Short hashes as ids for better long term caching.
+  config.optimization.merge({ moduleIds: 'hashed' })
+
   if (process.env.GRIDSOME_TEST) {
     config.output.pathinfo(true)
     config.optimization.minimize(false)
+    config.optimization.merge({ moduleIds: 'named' })
   }
 
   // helpes