+++ gridsome/lib/workers/__tests__/image-processor.spec.js
@@ -87,6 +87,14 @@ test('resize image', async () => {
   expect(height).toEqual(60)
 })
 
+test('rotate image based on EXIF', async () => {
+  const files = await process(['rotated.jpg'], { width: 480 })
+  const { width, height } = imageSize.sync(files[0].buffer)
+
+  expect(width).toEqual(480)
+  expect(height).toEqual(640)
+})
+
 test('crop image', async () => {
   const files = await process(['1000x600.png'], { width: 500, height: 500 })
   const { type, width, height } = imageSize.sync(files[0].buffer)
@@ -105,13 +113,49 @@ test('do not upscale images', async () => {
   expect(height).toEqual(250)
 })
 
+test('do not use processed image if file size gets bigger', async () => {
+  const files = await process([
+    'rotated.jpg',
+    '1000x600.jpg',
+    '350x250.png',
+    '1000x600-compressed.jpg',
+    '350x250-compressed.png'
+  ], { quality: 100 })
+
+  for (const file of files) {
+    const originalImage = fs.statSync(file.filePath)
+    const resultImage = fs.statSync(file.destPath)
+    expect(resultImage.size).toBeLessThanOrEqual(originalImage.size)
+  }
+})
+
+test('do not increase file size when compression is disabled', async () => {
+  const files = await process([
+    'rotated.jpg',
+    '1000x600.jpg',
+    '350x250.png',
+    '1000x600-compressed.jpg',
+    '350x250-compressed.png'
+  ], {}, {
+    images: {
+      compress: false
+    }
+  })
+
+  for (const file of files) {
+    const originalImage = fs.statSync(file.filePath)
+    const resultImage = fs.statSync(file.destPath)
+    expect(resultImage.size).toBeLessThanOrEqual(originalImage.size)
+  }
+})
+
 test('use cached process results', async () => {
   await fs.copy(
     path.join(context, 'assets', '1000x600.png'),
     path.join(imageCacheDir, '1000x600.97c148e.test.png')
   )
 
-  const files = await process(['1000x600.png'], { width: 1000 }, true)
+  const files = await process(['1000x600.png'], { width: 1000 }, { withCache: true })
   const cacheStats = await fs.stat(files[0].cachePath)
   const destStats = await fs.stat(files[0].destPath)
 
@@ -123,7 +167,14 @@ test('use cached process results', async () => {
   expect(destStats.size).toEqual(cacheStats.size)
 })
 
-async function process (filenames, options = {}, withCache = false) {
+async function process (
+  filenames,
+  options = {},
+  {
+    withCache = false,
+    images = {}
+  } = {}
+) {
   const config = {
     pathPrefix,
     imagesDir,
@@ -131,7 +182,10 @@ async function process (filenames, options = {}, withCache = false) {
     maxImageWidth: 1000,
     imageExtensions: ['.jpg', '.png', '.svg', '.gif', '.webp'],
     images: {
-      defaultBlur: 20
+      process: true,
+      defaultBlur: 20,
+      defaultQuality: 75,
+      ...images
     }
   }
 
@@ -145,7 +199,8 @@ async function process (filenames, options = {}, withCache = false) {
   await processImages({
     queue: processQueue.images.queue,
     cacheDir: withCache ? imageCacheDir : false,
-    outputDir: context
+    imagesConfig: config.images,
+    context
   })
 
   return Promise.all(assets.map(async ({ filePath, src, hash }) => {