+++ playground/gridsome.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  siteName: 'Gridsome Playground',
+  templates: {
+    Post: '/blog/:slug'
+  }
+}