+++ packages/plugin-google-analytics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/gridsome/gridsome/tree/master/packages/plugin-google-analytics/compare/@gridsome/plugin-google-analytics@0.1.0...@gridsome/plugin-google-analytics@0.1.1) (2020-05-26)
+
+**Note:** Version bump only for package @gridsome/plugin-google-analytics
+
+
+
+
+
 <a name="0.1.0"></a>
 # 0.1.0 (2019-01-10)